requirements.txt
fastmcp==2.9.0
python-jobspy>=1.1.82
requests>=2.32.3
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0,<2.10.0
pandas>=2.1.0
beautifulsoup4>=4.12.2
numpy>=1.26.0